Moved updateAction to utils
